requirements.txt
wxPython==4.2.2
pystray==0.19.5
pillow==11.1.0
winsdk==1.0.0b10
requests==2.32.3
pywin32==308
psutil==6.1.1
pynvml==12.0.0
pynput==1.8.1
pylnk3==0.4.3
win10toast==0.9

pywinstyles==1.8
colour==0.1.5